Add auto-refreshing and load JSON data in
This update auto refreshes commands for each guild it's in when the bot starts.
Also, I removed the guild ID parameter, as commands are refreshed with an ID provided at change time, based on other factors.
Finally, I remade the config files, allowing them to be better structured for future changes.